hash_passwordを共通化
