+++ spec/factories/members.rb
@@ -1,19 +1,6 @@
-def hash_password(key, salt = "")
-  return nil unless key.is_a? String
+require 'sinatra/crypt_helpers'
 
-  crypt_binname = case RUBY_PLATFORM
-    when /darwin/;  "crypt_darwin_amd64"
-    when /freebsd/; "crypt_freebsd_amd64"
-    when /linux/;   "crypt_linux_amd64"
-  end
-  path = File.expand_path("../../../ext/#{crypt_binname}", __FILE__)
-  hash, status = Open3.capture2(path, key, salt)
-  if status.exitstatus.zero?
-    hash.strip
-  else
-    nil
-  end
-end
+include Sinatra::CryptHelpers
 
 FactoryBot.define do
   factory :member do
